rename `docs` to `documentation`
