+++ rules/no-unreadable-array-destructuring.js
@@ -1,5 +1,5 @@
 'use strict';
-const getDocumentsUrl = require('./utils/get-documents-url');
+const getDocumentationUrl = require('./utils/get-documentation-url');
 
 const isCommaFollowedWithComma = (element, index, array) => {
 	return element === null && array[index + 1] === null;
@@ -28,7 +28,7 @@ module.exports = {
 	meta: {
 		type: 'suggestion',
 		docs: {
-			url: getDocumentsUrl(__filename)
+			url: getDocumentationUrl(__filename)
 		}
 	}
 };
