+++ rules/no-unused-properties.js
@@ -1,5 +1,5 @@
 'use strict';
-const getDocumentsUrl = require('./utils/get-documents-url');
+const getDocumentationUrl = require('./utils/get-documentation-url');
 
 const getDeclaratorOrPropertyValue = declaratorOrProperty => {
 	return declaratorOrProperty.init || declaratorOrProperty.value;
@@ -234,7 +234,7 @@ module.exports = {
 	meta: {
 		type: 'suggestion',
 		docs: {
-			url: getDocumentsUrl(__filename)
+			url: getDocumentationUrl(__filename)
 		}
 	}
 };
