+++ rules/regex-shorthand.js
@@ -1,6 +1,6 @@
 'use strict';
 const cleanRegexp = require('clean-regexp');
-const getDocumentsUrl = require('./utils/get-documents-url');
+const getDocumentationUrl = require('./utils/get-documentation-url');
 const quoteString = require('./utils/quote-string');
 
 const message = 'Use regex shorthands to improve readability.';
@@ -72,7 +72,7 @@ module.exports = {
 	meta: {
 		type: 'suggestion',
 		docs: {
-			url: getDocumentsUrl(__filename)
+			url: getDocumentationUrl(__filename)
 		},
 		fixable: 'code'
 	}
