+++ test/get-documentation-url.js
@@ -1,13 +1,13 @@
 import test from 'ava';
 import pkg from '../package.json';
-import getDocumentsUrl from '../rules/utils/get-documents-url';
+import getDocumentationUrl from '../rules/utils/get-documentation-url';
 
 test('returns the URL of the a named rule\'s documentation', t => {
 	const url = `https://github.com/sindresorhus/eslint-plugin-unicorn/blob/v${pkg.version}/docs/rules/foo.md`;
-	t.is(getDocumentsUrl('foo.js'), url);
+	t.is(getDocumentationUrl('foo.js'), url);
 });
 
 test('determines the rule name from the file', t => {
-	const url = `https://github.com/sindresorhus/eslint-plugin-unicorn/blob/v${pkg.version}/docs/rules/get-documents-url.md`;
-	t.is(getDocumentsUrl(__filename), url);
+	const url = `https://github.com/sindresorhus/eslint-plugin-unicorn/blob/v${pkg.version}/docs/rules/get-documentation-url.md`;
+	t.is(getDocumentationUrl(__filename), url);
 });
